Add annotation to generated classes

+++ src/org/fluentness/common/Utils.java
@@ -14,10 +14,12 @@ public class Utils {
                 }
             }
         }
-        if (!file.delete()) {
-            Logger.warning(this.getClass(), "Cannot delete %s", file.getPath());
-        } else {
-            Logger.info(this.getClass(), "Deleted file %s", file.getPath());
+        if (file.exists()) {
+            if (!file.delete()) {
+                Logger.warning(Utils.class, "Cannot delete %s", file.getPath());
+            } else {
+                Logger.info(Utils.class, "Deleted file %s", file.getPath());
+            }
         }
     }
 }
